VAM udated to use the performance optimised version of the experience editor
