+++ package.json
@@ -16,8 +16,8 @@
   "devDependencies": {
     "@babel/core": "^7.1.2",
     "@babel/preset-env": "^7.1.0",
-    "@fesk/scripts": "1.2.0-next.6f875fb9",
-    "@fesk/webpack-config": "1.2.0-next.6f875fb9",
+    "@fesk/scripts": "1.2.1",
+    "@fesk/webpack-config": "1.2.1",
     "babel-core": "^7.0.0-bridge",
     "babel-jest": "^23.6.0",
     "enzyme": "^3.7.0",
@@ -31,7 +31,8 @@
     "@canvas-panel/patchwork-plugin": "^0.1.1",
     "@canvas-panel/slideshow": "^0.1.1",
     "@fesk/react-bem": "^1.1.2",
-    "@iiif-mec/core": "^1.0.29",
+    "@iiif-mec/core": "1.2.2-19",
+    "@iiif-mec/experience-editor": "1.2.2-19",
     "@material-ui/core": "^3.6.1",
     "@material-ui/icons": "^3.0.1",
     "classnames": "^2.2.6",
@@ -41,13 +42,14 @@
     "langs": "^2.0.0",
     "notistack": "^0.5.1",
     "prop-types": "^15.6.2",
-    "react": "16.8.2",
+    "react": "16.8.6",
     "react-beautiful-dnd": "^10.0.2",
     "react-container-dimensions": "^1.4.1",
-    "react-dom": "16.8.2",
+    "react-dom": "16.8.6",
     "react-dropzone": "7.0.1",
     "react-input-range": "^1.3.0",
     "react-lazy-images": "^1.1.0",
+    "react-quill": "^1.3.3",
     "react-rnd": "^9.0.4",
     "react-scroll-wheel-handler": "^1.0.6",
     "react-virtualized": "^9.21.0",
