--- src/components/Preview/suggested.css
@@ -1,5 +1,0 @@
-/* to enable the bold in the patchwork annotation */ 
-.page-outer .patchwork-annotation-detail strong,
-.page-outer .patchwork-annotation-detail b {
-    font-weight: bold;
-}
