+++ src/defaults/index.js
@@ -3,9 +3,11 @@ import slideshow from './slideshow.js';
 import iiifManifest from './default.js';
 
 let rootManifestUrl;
+let rootCollection;
 
 try {
-  rootManifestUrl = process.env.COLLECTION_SERVER;
+  rootManifestUrl = process.env.ROOT_MANIFEST_URL || '';
+  rootCollection = process.env.COLLECTION_SERVER || '';
 } catch (ex) {
   // fallback settings
   const isLocalhost = () =>
@@ -16,10 +18,12 @@ try {
   rootManifestUrl = isLocalhost()
     ? 'http://localhost:8181/p3/'
     : 'https://iiif-collection.ch.digtest.co.uk/p3/';
+  rootCollection = rootManifestUrl;
 }
 
 export default {
   rootManifestUrl,
+  rootCollection,
   'annotated-zoom': annotatedZoom,
   slideshow,
   [`default`]: iiifManifest,
