+++ src/nunjucks/index.twig
@@ -2,7 +2,7 @@
 title: V&A Manifest Editor
 ---
 <style>
-  	html,body, app {
+  html,body, #app {
 		width: 100%;
 		height: 100%;
 		margin: 0;
@@ -25,39 +25,9 @@ title: V&A Manifest Editor
 		width: 100%;
 		height: 100%;
 	}
-	@keyframes spin {
-		from {
-			transform:rotate(0deg);
-		}
-		to {
-			transform:rotate(360deg);
-		}
-	}
-	.loading-indicator>svg>circle {
-		animation-name: spin;
-		animation-duration: 4000ms;
-		animation-iteration-count: infinite;
-		animation-timing-function: linear;
-		transform-origin: 50% 50%;
-	}
 </style>
 <div id="app">
 	<div class="loading-indicator">
-		<!-- for some resason animation stops whils the code evaluating, svg 
-			width="100" 
-			height="100" 
-			xmlns="http://www.w3.org/2000/svg" 
-			viewBox="0 0 100 100" 
-			preserveAspectRatio="xMidYMid" 
-			class="lds-rolling"
-		>
-			<circle 
-				cx="50"
-				cy="50"
-				fill="none" 
-				stroke="#ff9a00" stroke="#4658ac" stroke-width="10" r="35" stroke-dasharray="164.93361431346415 56.97787143782138" >
-			</circle>
-		</svg-->
 		<p>Loading Editor...</p>
 	<div>
 </div>
